Move common adversarial walker code to base class
